requirements.txt
pandas
numpy<2
scikit-surprise
requests
